rename thigns in prototype
